Fix creation of hdfs and httpfs users in pre script (SOFTWARE-608)
For hdfs, the "useradd" statement (adding the hdfs user) has the arguments "-G hadoop" to add the hdfs user to the hadoop group.
That fails (on el5) if the hadoop group does not exist at the time the scriptlet is run.
Moved that line to %post. Did the same thing for httpfs, which would have had a similar problem if the useradd statement didn't have a mistake in it (-G httpfs instead of -G hadoop).


git-svn-id: file:///p/condor/workspaces/vdt/svn/native/redhat/upcoming/hadoop@16925 4e558342-562e-0410-864c-e07659590f8c