drop broken _nocream %{?suffix} in el7 (SOFTWARE-2128)
I don't know if %suffix is a reserved macro name in rpm/rpmbuild or what,
but as evidenced by all our el7 builds, the %suffix macro expands to an
empty string, even when it's expected to be defined to _nocream .

We can get it working by changing this macro name to %release_suffix or
something like that, but to add it now might be confusing given that the
ones we ship do not have the suffix.

By the way, my main motivation for removing the broken stuff it is that
I'll be pushing this osg el7 section upstream (#5423).

git-svn-id: file:///p/condor/workspaces/vdt/svn/native/redhat/trunk/condor@22196 4e558342-562e-0410-864c-e07659590f8c