fix maven repo name for ivy (SOFTWARE-2161)
So, apparently a maven repo name ending with a '/' is a No-No.
At least for http://repo1.maven.org/maven2/ ...

Weirdly, http://repo1.maven.org/maven2//org/apache/ivy/ivy/2.2.0/ gives
a valid file listing, but any of the files listed under it (namely
http://repo1.maven.org/maven2//org/apache/ivy/ivy/2.2.0/ivy-2.2.0.jar)
give a 404 error, and the build fails.

Remove the double-slash and..voila!  But looking at the surrounding
code, the trailing slash kind of stands out anyway.

git-svn-id: file:///p/condor/workspaces/vdt/svn/native/redhat/branches/software-2161/hadoop@22492 4e558342-562e-0410-864c-e07659590f8c